timetable_app/requirements.txt
arrow==0.14.7
asgiref==3.3.1
Django==3.1.7
gunicorn==20.1.0
ics==0.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
TatSu==5.6.1
